# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Authentication and authorization REST API (JWT access + refresh, OTP, RBAC, registration invites).",
        "title": "Auth service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/v1/admin/registration-invites": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Create registration invite",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Required for human access tokens; omitted for verified service tokens",
                        "name": "X-CSRF-Token",
                        "in": "header"
                    },
                    {
                        "description": "Optional locked email and TTL",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httptransport.CreateRegistrationInviteRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/httptransport.CreateRegistrationInviteResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httptransport.ErrEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httptransport.ErrEnvelope"
                        }
                    },
                    "403": {
                        "description": "Not a superuser",
                        "schema": {
                            "$ref": "#/definitions/httptransport.ErrEnvelope"
                        }
                    }
                }
            }
        },
        "/v1/admin/service-accounts": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Create service account",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Required for human access tokens; omitted for verified service tokens",
                        "name": "X-CSRF-Token",
                        "in": "header"
                    },
                    {
                        "description": "Service account credentials and privilege",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httptransport.CreateServiceAccountRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/httptransport.CreateServiceAccountResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httptransport.ErrEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httptransport.ErrEnvelope"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/httptransport.ErrEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httptransport.ErrEnvelope"
                        }
                    }
                }
            }
        },
        "/v1/admin/signing-keys/rotate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Rotate signing key",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Required for human access tokens; omitted for verified service tokens",
                        "name": "X-CSRF-Token",
                        "in": "header"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/httptransport.ErrEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httptransport.ErrEnvelope"
                        }
                    }
                }
            }
        },
        "/v1/admin/users": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "List users",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Case-insensitive login or email substring",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Opaque keyset cursor; omit on first request",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page (max 100)",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httptransport.AdminUsersResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/httptransport.ErrEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httptransport.ErrEnvelope"
                        }
                    }
                }
            }
        },
        "/v1/admin/users/{userID}/ban": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Ban user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User UUID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Required for human access tokens; omitted for verified service tokens",
                        "name": "X-CSRF-Token",
                        "in": "header"
                    },
                    {
                        "description": "Reason",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httptransport.BanUserRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httptransport.ErrEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httptransport.ErrEnvelope"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/httptransport.ErrEnvelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httptransport.ErrEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httptransport.ErrEnvelope"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Unban user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User UUID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Required for human access tokens; omitted for verified service tokens",
                        "name": "X-CSRF-Token",
                        "in": "header"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httptransport.ErrEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httptransport.ErrEnvelope"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/httptransport.ErrEnvelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httptransport.ErrEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httptransport.ErrEnvelope"
                        }
                    }
                }
            }
        },
        "/v1/auth/has-role": {
            "post": {
                "description": "Intended for another service authorizing a human caller. The user ID is derived only from the verified access token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Check a role using an access token from JSON",
                "parameters": [
                    {
                        "description": "Access token and role name",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httptransport.TokenHasRoleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httptransport.HasRoleResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httptransport.ErrEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httptransport.ErrEnvelope"
                        }
                    }
                }
            }
        },
        "/v1/auth/has-role-with-tag": {
            "post": {
                "description": "Intended for another service authorizing a human caller. Role and tag inheritance use the normal authorization rules.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Check a role tag using an access token from JSON",
                "parameters": [
                    {
                        "description": "Access token, role name, and tag",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httptransport.TokenHasRoleWithTagRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httptransport.HasRoleWithTagResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httptransport.ErrEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httptransport.ErrEnvelope"
                        }
                    }
                }
            }
        },
        "/v1/auth/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login (password step)",
                "parameters": [
                    {
                        "description": "Credentials",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httptransport.LoginRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "otp_sent indicates whether OTP email was sent",
                        "schema": {
                            "$ref": "#/definitions/httptransport.LoginOTPResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httptransport.ErrEnvelope"
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "schema": {
                            "$ref": "#/definitions/httptransport.ErrEnvelope"
                        }
                    },
                    "403": {
                        "description": "Password must be changed",
                        "schema": {
                            "$ref": "#/definitions/httptransport.LoginPasswordExpiredResponse"
                        }
                    },
                    "423": {
                        "description": "Account locked",
                        "schema": {
                            "$ref": "#/definitions/httptransport.ErrEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httptransport.ErrEnvelope"
                        }
                    }
                }
            }
        },
        "/v1/auth/login/magic-link": {
            "post": {
                "description": "Public passwordless flow. Always returns 200 to avoid enumeration; a link is emailed only when the login exists and has an email.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Request a magic login link",
                "parameters": [
                    {
                        "description": "Login to send a link to",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httptransport.MagicLinkStartRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httptransport.MagicLinkStartResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httptransport.ErrEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httptransport.ErrEnvelope"
                        }
                    }
                }
            }
        },
        "/v1/auth/login/magic-link/verify": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Complete magic-link login",
                "parameters": [
                    {
                        "description": "One-time token and device binding",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httptransport.MagicLinkVerifyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Includes csrf_token",
                        "schema": {
                            "$ref": "#/definitions/httptransport.TokenPairResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httptransport.ErrEnvelope"
                        }
                    },
                    "401": {
                        "description": "Invalid or expired link",
                        "schema": {
                            "$ref": "#/definitions/httptransport.ErrEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httptransport.ErrEnvelope"
                        }
                    }
                }
            }
        },
...
        "/v1/auth/logout": {
            "post": {
                "description": "Revokes the session named by refresh_token in the body (multi-account, non-ambient) or, if absent, the HttpOnly cookie session (requires X-CSRF-Token).",
                "tags": [
                    "auth"
                ],
                "summary": "Logout",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Required only for the cookie path",
                        "name": "X-CSRF-Token",
                        "in": "header"
                    },
                    {
                        "description": "Optional refresh_token to revoke a specific account",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/httptransport.LogoutRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "403": {
                        "description": "CSRF validation failed",
                        "schema": {
                            "$ref": "#/definitions/httptransport.ErrEnvelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "httptransport.CreateRegistrationInviteRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "superuser": {
                    "description": "Superuser grants superuser access to the account registered with this invite.",
                    "type": "boolean"
                },
                "ttl_seconds": {
                    "type": "integer"
                }
            }
        },
        "httptransport.CreateRegistrationInviteResponse": {
            "type": "object",
            "properties": {
                "expires_at": {
                    "type": "string"
                },
                "registration_url": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
...
        "httptransport.CreateServiceAccountRequest": {
            "type": "object",
            "properties": {
                "login": {
                    "type": "string"
                },
                "secret": {
                    "type": "string"
                },
                "superuser": {
                    "type": "boolean"
                }
            }
        },
...
        "httptransport.ErrEnvelope": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and the access JWT.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}